fix: rename command name
